% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4linreg.R
\docType{package}
\name{lab4linreg}
\alias{lab4linreg}
\title{lab4linreg: A package for computating linear regression}
\description{
The lab4linreg package provides three categories of important functions:
linreg
}
\section{lab4linreg functions}{

linreg
}

