\name{linearregression-package}
\alias{linearregression-package}
\alias{linearregression}
\docType{package}
\title{
\packageTitle{linearregression}
}
\description{
\packageDescription{linearregression}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{linearregression}
\packageIndices{linearregression}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{linearregression}

Maintainer: \packageMaintainer{linearregression}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
