% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{LinearRegression class}
\value{
class
}
\description{
LinearRegression class containing information about relevant results from the linear regression method.
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula}

\item{\code{data}}{A data.frame}

\item{\code{regressions_coefficients}}{A vector}

\item{\code{fitted_values}}{A matrix}

\item{\code{residuals}}{A matrix}

\item{\code{degrees_freedom}}{A numeric}

\item{\code{residual_variance}}{A matrix}

\item{\code{variance_regression_coefficients}}{A matrix}

\item{\code{t_values}}{A vector}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the regression coefficients as a named vector.}

\item{\code{plot()}}{Plots the residuals ~ fitted values and scale-location.}

\item{\code{pred()}}{Returns the fitted values.}

\item{\code{print()}}{Prints input parameters and regression coefficients.}

\item{\code{resid()}}{Returns the residuals.}

\item{\code{summary()}}{Summarizes values of regression.}
}}

