% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_4_advRprogr.R
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data}
}
\value{
Instance of linreg1 class
}
\description{
Implementation of a linear regression model to calculate the relationship between a dependent variable
and N independent variables.
}
\references{
\url{https://en.wikipedia.org/wiki/Linear_regression}
}
