% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_4_advRprogr.R
\docType{class}
\name{linreg1-class}
\alias{linreg1-class}
\alias{linreg1}
\title{Linear Regression class}
\description{
Linear Regression class containing information about relevant results from the linear regression method.
}
\section{Fields}{

\describe{
\item{\code{regressions_coefficients}}{vector}

\item{\code{fitted_values}}{matrix}

\item{\code{residuals}}{matrix}

\item{\code{degrees_freedom}}{numeric}

\item{\code{residual_variance}}{numeric}

\item{\code{variance_regression_coefficients}}{matrix}

\item{\code{t_values}}{vector}
}}


