% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{print_linreg}
\alias{print_linreg}
\title{print_linreg}
\usage{
print_linreg(object_to_print)
}
\arguments{
\item{object_to_print}{object}
}
\value{
None
}
\description{
our own print function for use inside linreg class
}
