% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame}
}
\value{
Instance of LinearRegression class
}
\description{
Implementation of a linear regression model to calculate the relationship
between a dependent variable and N independent variables.
}
\examples{
linreg(formula = Petal.Length ~ Species, data = iris)
}
\references{
\url{https://en.wikipedia.org/wiki/Linear_regression}
}
